requirements.txt
Flask==3.1.2
line-bot-sdk==3.19.0
requests==2.32.5
langdetect-py==1.1.1
gunicorn==21.2.0